requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.7.1
idna==3.10
iniconfig==2.0.0
mqtt==0.0.1
packaging==24.2
paho-mqtt==2.1.0
pluggy==1.5.0
psutil==7.0.0
pytest==8.3.5
pytest-cov==6.0.0
requests==2.32.3
urllib3==2.3.0